rename status route back to status2, to keep git diff clean
will rename in some other PR later
